Plot page needed to be reworked a bit to allow nextjs to prerender the page without loading all the dynamic parts, but now works in run dev and run build. Seems to be some sort of lag when first loading the download and plot pages that might be related to the disclaimer popup.
